new util package, optimized communication to the user, user Messages centralized
For each command the user types he gets an answer if it was valid or
not. the new help command shows him all possible options, welcome
message comes now only when starting the program. The new util package
includes all classes which help to structure the programm but are not
critical. E.g. the centralized messages or the validation class.