workingDirectoryHandler.py added, commandLineHand
workingDirectoryHandler.py added, which contains two functions

determinWorkingDirectory():
checks if a path to a working directory is saved in
util.workingDirectory.txt, if not asks the user to enter path, the
checks if path is valid [has a fatherDirectory(via function
fatherDirectoryExists()) and is not existend already existend] until
user made a proper insertion, after that he saves the path in
util.workingDirectory.txt.

fatherDirectoryExists():
checks if fatherDirectory exists

function receiveAndValidateInput() replaced by userCommunication in
commandLineHandler, because of improved naming. This function at first
executes printStatusOfExecution() and then runs
receiveAndValidateInput(), which now does exactly what it is named